requirements.txt
joblib_progress
joblib
pdf2image
rich
opencv-python
pillow
pymupdf